Reorganize the code in three modules : coreserver-api, core-server and wpsservice.
