--- core-server/src/main/java/org/orbisgis/orbisserver/coreserver/model/Service.java
@@ -1,80 +1,0 @@
-/*
- * OrbisServer is an OSGI web application to expose OGC services.
- *
- * OrbisServer is part of the OrbisGIS platform
- *
- * OrbisGIS is a java GIS application dedicated to research in GIScience.
- * OrbisGIS is developed by the GIS group of the DECIDE team of the
- * Lab-STICC CNRS laboratory, see <http://www.lab-sticc.fr/>.
- *
- * The GIS group of the DECIDE team is located at :
- *
- * Laboratoire Lab-STICC – CNRS UMR 6285
- * Equipe DECIDE
- * UNIVERSITÉ DE BRETAGNE-SUD
- * Institut Universitaire de Technologie de Vannes
- * 8, Rue Montaigne - BP 561 56017 Vannes Cedex
- *
- * OrbisServer is distributed under LGPL 3 license.
- *
- * Copyright (C) 2017 CNRS (Lab-STICC UMR CNRS 6285)
- *
- *
- * OrbisServer is free software: you can redistribute it and/or modify it under the
- * terms of the GNU Lesser General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any later
- * version.
- *
- * OrbisServer is distributed in the hope that it will be useful, but WITHOUT ANY
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
- * A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License along with
- * OrbisServer. If not, see <http://www.gnu.org/licenses/>.
- *
- * For more information, please consult: <http://www.orbisgis.org/>
- * or contact directly:
- * info_at_ orbisgis.org
- */
-package org.orbisgis.orbisserver.coreserver.model;
-
-import java.util.List;
-import java.util.Map;
-
-/**
- * Interface to define a service which can be plugged to the server
- *
- * @author Sylvain PALOMINOS
- */
-public interface Service {
-
-    /**
-     * Execute an operation with the given data
-     *
-     * @param request Request containing all the data for the execution
-     *
-     * @return A statusInfo object containing all the information about the execution
-     */
-    public StatusInfo executeOperation(ExecuteRequest request);
-
-    /**
-     * Gets the status of an execution
-     *
-     * @param request Request containing the identifier of an execution
-     *
-     * @return A statusInfo object containing all the information about the execution
-     */
-    public StatusInfo getStatus(StatusRequest request);
-
-    public Result getResult(StatusRequest request);
-
-    public List<Operation> getAllOperation();
-
-    boolean hasOperation(String id);
-
-    Operation getOperation(String id);
-
-    void shutdown();
-
-    void start(Map<String, Object> propertyMap);
-}
