+++ core-server/src/main/java/org/orbisgis/orbisserver/coreserver/model/Session.java
@@ -42,6 +42,9 @@ import org.apache.commons.io.filefilter.NameFileFilter;
 import org.h2gis.utilities.JDBCUtilities;
 import org.h2gis.utilities.SFSUtilities;
 import org.h2gis.utilities.TableLocation;
+import org.orbisgis.orbisserver.api.model.*;
+import org.orbisgis.orbisserver.api.service.Service;
+import org.orbisgis.orbisserver.api.service.ServiceFactory;
 import org.orbisgis.orbisserver.coreserver.web.MainController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
