+++ coreserver-api/src/main/java/org/orbisgis/orbisserver/api/CoreServerController.java
@@ -1,0 +1,11 @@
+package org.orbisgis.orbisserver.api;
+
+import org.orbisgis.orbisserver.api.service.ServiceFactory;
+
+/**
+ * @author Sylvain PALOMINOS
+ */
+public interface CoreServerController{
+
+    void addServiceFactory(ServiceFactory serviceFactory);
+}
