+++ wpsservice/pom.xml
@@ -14,15 +14,86 @@
     <artifactId>wpsservice</artifactId>
     <version>1.0-SNAPSHOT</version>
 
-    <name>core-server</name>
+    <name>wpsservice</name>
 
     <packaging>wisdom</packaging>
 
     <properties>
-            <wisdom.version>0.10.0</wisdom.version>
+        <wisdom.version>0.10.0</wisdom.version>
     </properties>
 
     <dependencies>
+        <dependency>
+            <groupId>org.wisdom-framework</groupId>
+            <artifactId>wisdom-api</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-databind</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.felix</groupId>
+            <artifactId>org.apache.felix.ipojo.annotations</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <scope>provided</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>com.googlecode.gettext-commons</groupId>
+            <artifactId>gettext-commons</artifactId>
+            <version>${gettext-commons-version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.webjars</groupId>
+            <artifactId>bootstrap</artifactId>
+            <version>3.1.1</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.webjars</groupId>
+            <artifactId>jquery</artifactId>
+            <version>2.1.0-2</version>
+        </dependency>
+
+        <!-- Test dependencies -->
+        <dependency>
+            <groupId>org.wisdom-framework</groupId>
+            <artifactId>wisdom-test</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.assertj</groupId>
+            <artifactId>assertj-core</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-all</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.fluentlenium</groupId>
+            <artifactId>fluentlenium-assertj</artifactId>
+        </dependency>
+        <dependency>
+            <!-- slf4j binding used for tests -->
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-jcl</artifactId>
+            <version>1.7.6</version>
+            <scope>test</scope>
+        </dependency>
+
+
+
+
         <dependency>
             <groupId>org.orbiswps</groupId>
             <artifactId>server</artifactId>
@@ -62,13 +133,13 @@
         </dependency>
 
         <dependency>
-            <groupId>org.javassist</groupId>
-            <artifactId>javassist</artifactId>
-            <version>3.19.0-GA</version>
+            <groupId>org.wisdom-framework</groupId>
+            <artifactId>wisdom-vertx-engine</artifactId>
+            <version>0.10.0</version>
         </dependency>
         <dependency>
             <groupId>org.orbisgis.orbisserver</groupId>
-            <artifactId>coreserver</artifactId>
+            <artifactId>coreserver-api</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
 
@@ -115,6 +186,26 @@
                         </configuration>
                     </execution>
                 </executions>
+            </plugin>
+            <plugin>
+                <groupId>com.googlecode.gettext-commons</groupId>
+                <version>${gettext-maven-plugin-version}</version>
+                <executions>
+                    <execution>
+                        <id>convert-po-class</id>
+                        <phase>compile</phase>
+                        <goals>
+                            <goal>dist</goal>
+                        </goals>
+                    </execution>
+                </executions>
+                <artifactId>gettext-maven-plugin</artifactId>
+                <configuration>
+                    <poDirectory>src/main/resources/org/orbisgis/orbisserver/wpsservice/translation/</poDirectory>
+                    <targetBundle>org.orbisgis.orbisserver.wpsservice.Messages</targetBundle>
+                    <keywords>-ktr -kmarktr</keywords>
+                    <outputFormat>properties</outputFormat>
+                </configuration>
             </plugin>
         </plugins>
     </build>
